Fix 4 remaining bugs: Thumb32 branch, findExecSection, shared rodata, segment test
- Thumb32 branch encoding: use int64 arithmetic to avoid uint16 truncation
  of sign-extended negative offsets
- findExecSection: only return .text if target address is within it,
  instead of returning unconditionally
- String encryption: track already-encrypted strings across functions
  to prevent double-encrypting shared .rodata entries
- TestMakeSegmentsWritable: align test expectations with actual behavior
  (non-exec segments get PF_W, exec segments skip to avoid W+X)

Also confirmed two reported issues were false positives:
- ROR operand count: OpSRor correctly takes 3 stack operands
- RTLR uint64: table is only generated for ARM64 path where uint64 is correct

+++ pkg/binary/elf/trampoline.go
@@ -158,7 +158,7 @@ func writeThumb32MovT(w io.Writer, rd int, imm16 uint32) {
 }
 
 func writeThumb32BranchW(w io.Writer, offset int32) {
-	imm := offset >> 1
+	imm := int64(offset) >> 1
 	s := uint16(0)
 	if imm < 0 { s = 1 }
 	imm10 := uint16((imm >> 11) & 0x3FF)
